+++ changelog.js
@@ -1,4 +1,9 @@
 export const changelog = {
+  'v1.3.2': `### [v1.3.2] - 11-05-2024
+  - CORRECTIF - Ajout d'une config NPM_APP_DIR afin de récupérer le chemin dans lequel l'app est installé afin de récupérer le fichier .jar
+  - CORRECTIF - Dans la fonction Log, correction d'une erreur avec value.length
+  `,
+
   'v1.3.1': `### [v1.3.1] - 11-05-2024
   - CORRECTIF - Ajout d'un fichier .jar pour extraire la BDD du PDA avec la commande run -e
   `,